move smallintmap to oldsmallintmap
