+++ src/libstd/std.rc
@@ -82,7 +82,7 @@ pub mod list;
 pub mod map;
 pub mod priority_queue;
 pub mod rope;
-pub mod smallintmap;
+pub mod oldsmallintmap;
 pub mod sort;
 pub mod treemap;
